feat: list planes with ascending id
The device has a list of all available planes. These are sorted depending on
their type and zpos value. But there are subsets without ordering depending
on that. Order these by their id implicitly in an ascending manner.

For example plane P1 with same type and zpos but higher id as some plane P2
will be after P2 in the list. In the past it was the other way around.